Refactor: WinningNumbers 클래스 삭제
WinningNumbers 삭제 후 Lotto 클래스로 통합
그에 따른 의존 클래스 수정 및 테스트 수정
