+++ src/test/java/model/DrawResultsTest.java
@@ -23,7 +23,7 @@ class DrawResultsTest {
 
     @Test
     void 구매한_로또가_전부_낙첨되면_수익률은_손해이다() {
-        WinningNumbers winningNumbers = WinningNumbers.create(List.of(7, 8, 9, 10, 11, 12));
+        Lotto winningNumbers = Lotto.create(List.of(7, 8, 9, 10, 11, 12));
         drawResults.calculateResults(lottos, winningNumbers);
         double profit = drawResults.calculateProfit(purchaseAmount);
         assertThat(profit).isLessThan(1);
@@ -31,7 +31,7 @@ class DrawResultsTest {
 
     @Test
     void 구매한_로또가_전부_당첨되면_수익률은_이익이다() {
-        WinningNumbers winningNumbers = WinningNumbers.create(List.of(1, 2, 3, 4, 5, 6));
+        Lotto winningNumbers = Lotto.create(List.of(1, 2, 3, 4, 5, 6));
         drawResults.calculateResults(lottos, winningNumbers);
         double profit = drawResults.calculateProfit(purchaseAmount);
         assertThat(profit).isGreaterThan(1);
