+++ src/test/java/model/RankingTest.java
@@ -1,10 +1,8 @@
 package model;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 class RankingTest {
